GTDB-tk fix: use batchfile from updated dastools output, declare output specifically, merge archaea and bacteria output into the same file.

+++ modules/binning/gtdbtk.nf
@@ -1,5 +1,5 @@
 process gtdbtk {
-        publishDir params.outdir + "/binning/gtdb", mode: params.publish_mode, pattern: 'gtdb_*'
+        publishDir params.outdir + "/binning/gtdb/", mode: params.publish_mode, pattern: 'gtdb_*'
         tag "GTDB-tk on $sample_id"
 
         input:
@@ -7,7 +7,7 @@ process gtdbtk {
         path(path_gtdb_tk)
 
         output:
-        tuple val(sample_id), path("gtdb_${sample_id}"), emit: gtdb_out
+        tuple val(sample_id), path("gtdb_${sample_id}.tsv"), emit: gtdb_out
         path "versions.yml", emit: versions
 
         shell:
@@ -15,12 +15,23 @@ process gtdbtk {
         export GTDBTK_DATA_PATH=${path_gtdb_tk}
         gtdbtk classify_wf --cpus ${task.cpus} --pplacer_cpus ${task.cpus} \
                 --skip_ani_screen \
-                --genome_dir ${dastool_bins} -x fa --out_dir gtdb_${sample_id}
-        
+                --batchfile ${dastool_bins} -x fa --out_dir gtdb_${sample_id}       
+	
+	if [[ ! -f gtdb_${sample_id}/gtdbtk.bac120.summary.tsv ]]
+        then
+                if [[ ! -f gtdb_${sample_id}/gtdbtk.ar53.summary.tsv ]]
+                then
+                	echo "Something went seriously wrong!"
+			exit
+		fi
+        fi
+        cat ./gtdb_${sample_id}/gtdbtk.*.tsv | head -n 1 > gtdb_${sample_id}.tsv
+        cat ./gtdb_${sample_id}/gtdbtk.*.tsv | grep -v -f gtdb_${sample_id}.tsv >> ./gtdb_${sample_id}.tsv
+	
         cat <<-END_VERSIONS > versions.yml
         "${task.process}":
             GTDB-tk: \$( gtdbtk -v | sed -e "s/gtdbtk: version //g" | sed -e "s/ Copyright.*//g" )
         END_VERSIONS
         """
-
 }
+
